+++ src/controller/storage.controller.ts
@@ -1,31 +1,6 @@
-import { GetObjectCommand, HeadObjectCommand, PutObjectCommand, S3Client } from "@aws-sdk/client-s3"
 import { Request, Response } from "express"
 import { CatchError, TryError } from "../utils/error"
-import { getSignedUrl } from "@aws-sdk/s3-request-presigner"
-import { SessionInterfece } from "../middleware/auth.middleware"
-
-const connection = new S3Client({
-    region: process.env.REGION,
-    endpoint: `https://s3-${process.env.REGION}.amazonaws.com`,
-    credentials: {
-        accessKeyId: process.env.AWS_ACCESS_KEY_ID!,
-        secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY!
-    }
-})
-
-const isFileExist = async (path: string) => {
-    try {
-        const command = new HeadObjectCommand({
-            Bucket: process.env.S3_BUCKET,
-            Key: path
-        })
-        await connection.send(command) 
-        return true
-        
-    } catch (err) {
-        return false
-    }
-}
+import { downloadObject, isFileExist, uploadObject } from "../utils/s3"
 
 export const downloadFile = async (req: Request, res: Response) => {
     try {
@@ -38,19 +13,11 @@ export const downloadFile = async (req: Request, res: Response) => {
         if(!isExist)
             throw TryError('File does not  exist', 404) 
 
-        const option = {
-            Bucket: process.env.S3_BUCKET,
-            Key: path
-        }
-
-        const command = new GetObjectCommand(option)
-
-        const url = await getSignedUrl(connection, command, {expiresIn: 60})
-        res.json({url}) 
-
+        const url = await downloadObject(path)
+        res.json({url})
+       
     } catch (err) {
         CatchError(err, res, "Failed to generate download url")
-        
     }
 }
 
@@ -62,12 +29,7 @@ export const uploadFile = async (req: Request, res: Response) => {
         if(!path || !type) 
             throw TryError("Invalid request path or type is required", 400)
 
-        const command = new PutObjectCommand({
-            Bucket: process.env.S3_BUCKET,
-            Key: path,
-            ContentType: type
-        })
-        const url = await getSignedUrl(connection, command, {expiresIn: 60})
+        const url = await uploadObject(path, type)
         res.json({url})
         
     } catch (err) {
