+++ src/utils/s3.ts
@@ -1,0 +1,47 @@
+import { HeadObjectCommand, PutObjectCommand, S3Client } from "@aws-sdk/client-s3"
+import { GetObjectCommand } from "@aws-sdk/client-s3"
+import { getSignedUrl } from "@aws-sdk/s3-request-presigner"
+
+export const connection = new S3Client({
+    region: process.env.REGION,
+    endpoint: `https://s3-${process.env.REGION}.amazonaws.com`,
+    credentials: {
+        accessKeyId: process.env.AWS_ACCESS_KEY_ID!,
+        secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY!
+    }
+})
+
+export const isFileExist = async (path: string) => {
+    try {
+        const command = new HeadObjectCommand({
+            Bucket: process.env.S3_BUCKET,
+            Key: path
+        })
+        await connection.send(command) 
+        return true
+        
+    } catch (err) {
+        return false
+    }
+}
+
+export const downloadObject = async (path: string, expiry: number = 60) => {
+    const option = {
+        Bucket: process.env.S3_BUCKET,
+        Key: path
+    }
+    const command = new GetObjectCommand(option)
+    const url = await getSignedUrl(connection, command, {expiresIn: expiry})
+    return url 
+}
+
+export const uploadObject = async (path: string, type: string) => {
+
+    const command = new PutObjectCommand({
+        Bucket: process.env.S3_BUCKET,
+        Key: path,
+        ContentType: type
+    })
+    const url = await getSignedUrl(connection, command, {expiresIn: 60})
+    return url
+}
